requirements.txt
numpy==2.0.1
# open3d==0.18.0+b271acb
open3d==0.18.0
opencv_python==4.10.0.84
pyrealsense2_macosx==2.53.1
